feat: optimistic update with revalidate false
