+++ examples/nextjs-todos/.wundergraph/operations/UpdateTodoOrder.graphql
@@ -1,0 +1,5 @@
+mutation ($id: Int, $order: db_IntFieldUpdateOperationsInput) {
+	db_updateOneTodo(where: { id: $id }, data: { order: $order }) {
+		id
+	}
+}
