+++ examples/nextjs-todos/.wundergraph/wundergraph.config.ts
@@ -6,6 +6,9 @@ import operations from './wundergraph.operations';
 const db = introspect.postgresql({
 	apiNamespace: 'db',
 	databaseURL: 'postgresql://admin:admin@localhost:54322/example?schema=public',
+	introspection: {
+		disableCache: true,
+	},
 });
 
 // configureWunderGraph emits the configuration
