+++ examples/nextjs-todos/package.json
@@ -18,6 +18,7 @@
   "dependencies": {
     "@wundergraph/nextjs": "^0.5.0",
     "@wundergraph/sdk": "^0.123.0",
+    "framer-motion": "^7.6.12",
     "graphql": "^16.3.0",
     "next": "^12.1.6",
     "react": "^18.0.0",
