+++ examples/nextjs-todos/pages/index.tsx
@@ -1,38 +1,55 @@
 import { NextPage } from 'next';
 import { withWunderGraph, useQuery, useMutation } from '../components/generated/nextjs';
 import React, { Fragment, useRef, useState } from 'react';
-import { DeleteTodoInput, EditTodoInput, UpdateCompleteTodoInput } from '../components/generated/models';
+import TodoItem from '../components/TodoItem';
+import NavBar from '../components/Navbar';
 import { mutate } from 'swr';
 
 const Home: NextPage = () => {
 	const [title, setTitle] = useState<string>('');
 	const titleRef = useRef<HTMLInputElement>();
-	const todos = useQuery({ operationName: 'Todos' });
+	const todos = useQuery({
+		operationName: 'Todos',
+	});
+
 	const createTodo = useMutation({ operationName: 'CreateTodo' });
 
 	async function addTodo() {
 		if (title.trim().length > 0) {
-			let addTodoData = { title: title };
-			await mutate({ operationName: 'Todos' }, async (todosItems) => {
-				const newTodo = await createTodo.trigger(addTodoData);
-				if (newTodo.db_createOneTodo) {
-					let addedItem = { id: newTodo.db_createOneTodo.id, title: addTodoData.title, completed: false };
-					todosItems.db_findManyTodo.push(addedItem);
-					cancelAdd();
-				}
-				return todosItems;
-			});
+			mutate(
+				{ operationName: 'Todos' },
+				async (todos) => {
+					if (todos) {
+						//make deep copy of todos
+						let newTodos = JSON.parse(JSON.stringify(todos));
+						let savedTodo = await createTodo.trigger({ title: title });
+						if (savedTodo.db_createOneTodo) {
+							let saveItem = {
+								id: savedTodo.db_createOneTodo.id,
+								title: title,
+								completed: false,
+								order: savedTodo.db_createOneTodo.id,
+							};
+							newTodos.db_findManyTodo.push(saveItem);
+						}
+						clearAdd();
+						return newTodos;
+					}
+				},
+				{ revalidate: false, rollbackOnError: true }
+			);
 		}
 	}
+
 	async function titleKeyHandler(event: React.KeyboardEvent<HTMLInputElement>) {
 		if (event.key === 'Escape') {
-			cancelAdd();
+			clearAdd();
 		} else if (event.key === 'Enter') {
 			await addTodo();
 		}
 	}
 
-	function cancelAdd() {
+	function clearAdd() {
 		setTitle('');
 		titleRef.current.blur();
 	}
@@ -81,199 +98,14 @@ const Home: NextPage = () => {
 						</div>
 					</div>
 					<div className={'absolute mt-2'}>
-						{todos?.data?.db_findManyTodo?.map((todo, index: number) => {
-							return (
-								<div key={todo.id}>
-									<TodoItem todo={todo} lastItem={index === todos.data.db_findManyTodo.length - 1} />
-								</div>
-							);
-						})}
+						{todos?.data?.db_findManyTodo?.map((todo, index: number) => (
+							<TodoItem key={todo.id} todo={todo} lastItem={index === todos.data.db_findManyTodo.length - 1} />
+						))}
 					</div>
 				</div>
 			</div>
 		</Fragment>
 	);
 };
-
-function TodoItem({ todo, lastItem }) {
-	const [currentTodo, setCurrentTodo] = useState(todo);
-	const [editMode, setEditMode] = useState<boolean>(false);
-
-	const updateCompleteTodo = useMutation({
-		operationName: 'UpdateCompleteTodo',
-	});
-	const deleteTodoOperation = useMutation({
-		operationName: 'DeleteTodo',
-	});
-	const updateTodo = useMutation({ operationName: 'EditTodo' });
-
-	async function updateCompletedStatus(e: React.ChangeEvent<HTMLInputElement>) {
-		let newCheckedStatus: boolean = e.target.checked;
-		let updateCompleteTodoStatus: UpdateCompleteTodoInput = {
-			id: currentTodo.id,
-			complete: {
-				set: newCheckedStatus,
-			},
-		};
-		await mutate({ operationName: 'Todos' }, async (todos) => {
-			const updatedCompleteTodoStatus = await updateCompleteTodo.trigger(updateCompleteTodoStatus);
-			let updatedTodos = todos.db_findManyTodo.map((todo) => {
-				if (todo.id === updatedCompleteTodoStatus.db_updateOneTodo.id) {
-					return updatedCompleteTodoStatus.db_updateOneTodo;
-				}
-				return todo;
-			});
-			return updatedTodos;
-		});
-	}
-
-	async function deleteTodo(id: number) {
-		const deleteTodoArg: DeleteTodoInput = { id: id };
-		await mutate({ operationName: 'Todos' }, async (todos) => {
-			const deletedTodo = await deleteTodoOperation.trigger(deleteTodoArg);
-			if (!deletedTodo.db_deleteOneTodo) {
-				return todos.db_findManyTodo;
-			}
-			let filteredTodos = todos.db_findManyTodo.filter((todo) => todo.id !== deletedTodo.db_deleteOneTodo.id);
-			return filteredTodos;
-		});
-	}
-
-	async function editTodo() {
-		if (currentTodo.title.trim().length > 0) {
-			let updateTodoTitle: EditTodoInput = {
-				id: currentTodo.id,
-				title: {
-					set: currentTodo.title,
-				},
-			};
-
-			await mutate({ operationName: 'Todos' }, async (todos) => {
-				const updatedTodo = await updateTodo.trigger(updateTodoTitle);
-				let updatedTodos = todos.db_findManyTodo.map((todo) => {
-					if (todo.id === updatedTodo.db_updateOneTodo.id) {
-						return updatedTodo.db_updateOneTodo;
-					}
-					return todo;
-				});
-				return updatedTodos;
-			});
-		}
-	}
-
-	function cancelEdit() {
-		setCurrentTodo({
-			...currentTodo,
-			title: todo.title,
-		});
-		setEditMode(false);
-	}
-
-	async function titleKeyHandler(event: React.KeyboardEvent<HTMLInputElement>) {
-		if (event.key === 'Escape') {
-			cancelEdit();
-		} else if (event.key === 'Enter') {
-			await editTodo();
-		}
-	}
-
-	function enableEditMode() {
-		setEditMode(true);
-	}
-
-	return (
-		<Fragment>
-			{!editMode && (
-				<div
-					className={`flex justify-between pt-4 pb-2 m-2 px-2 hover:px-3 w-72 hover:bg-zinc-600 hover:rounded
-					${!lastItem ? `border-solid border-0 border-b border-zinc-500` : ''}`}
-				>
-					<Fragment>
-						<div className="flex items-center mb-2">
-							<input
-								type="checkbox"
-								onChange={updateCompletedStatus}
-								checked={currentTodo.completed}
-								className={'h-4 w-4 rounded-full accent-pink-500'}
-							/>
-							<div
-								onClick={enableEditMode}
-								className={`cursor-pointer ml-3 text-sm font-medium text-gray-300  ${
-									currentTodo.completed ? 'line-through' : ''
-								}`}
-							>
-								<span className={'break-all'}>{todo.title}</span>
-							</div>
-						</div>
-						<div onClick={() => deleteTodo(todo.id)} className={'flex flex-col justify-start ml-2 cursor-pointer'}>
-							<svg
-								xmlns="http://www.w3.org/2000/svg"
-								fill="none"
-								viewBox="0 0 24 24"
-								strokeWidth={1.5}
-								stroke="#fafafa"
-								className="w-5 h-5 hover:bg-zinc-500 hover:rounded"
-							>
-								<path strokeLinecap="round" strokeLinejoin="round" d="M6 18L18 6M6 6l12 12" />
-							</svg>
-						</div>
-					</Fragment>
-				</div>
-			)}
-			{editMode && (
-				<Fragment>
-					<div className={'relative'}>
-						<input
-							type="text"
-							onKeyDown={titleKeyHandler}
-							value={currentTodo.title}
-							onChange={(e) => {
-								setCurrentTodo({
-									...currentTodo,
-									title: e.target.value,
-								});
-							}}
-							className={`${editMode ? 'mb-1' : ''}
-							${currentTodo.completed ? 'line-through' : ''}
-								border-solid border-0 border-b border-pink-400
-								py-3 pl-5 pr-10 w-72 bg-gray-900 text-white focus:outline-none 
-								`}
-						/>
-
-						<div onClick={editTodo} className={'absolute right-6 top-3 cursor-pointer hover:bg-zinc-500 hover:rounded'}>
-							<svg
-								xmlns="http://www.w3.org/2000/svg"
-								fill="none"
-								viewBox="0 0 24 24"
-								strokeWidth={1.5}
-								stroke="white"
-								className="w-6 h-6"
-							>
-								<path strokeLinecap="round" strokeLinejoin="round" d="M4.5 12.75l6 6 9-13.5" />
-							</svg>
-						</div>
-					</div>
-				</Fragment>
-			)}
-		</Fragment>
-	);
-}
-
-function NavBar() {
-	return (
-		<nav className="px-2 py-2.5 sm:px-4 bg-gray-900 border-solid border-0 border-b border-zinc-700">
-			<div>
-				<div className="flex items-center px-5 py-2">
-					<a href="https://wundergraph.com" target="_blank">
-						<img src="/wundergraph.svg" className="h-12 pr-3" alt="WunderGraph logo" />
-					</a>
-					<span className="self-center text-xl font-semibold whitespace-nowrap bg-gradient-to-r from-purple-400 to-pink-600 bg-clip-text text-transparent">
-						WunderGraph
-					</span>
-				</div>
-			</div>
-		</nav>
-	);
-}
 
 export default withWunderGraph(Home);
